refactor: rename survey test package to singular
Moves survey translation test files from the 'surveys' package to 'survey' to maintain naming consistency within the project structure.
